Implementing roles & database

+++ app.py
@@ -2,12 +2,13 @@ from flask import Flask, request, jsonify
 from database import db
 from models.user import User
 from flask_login import LoginManager, login_user, current_user,logout_user, login_required
+import bcrypt
 
 
 
 app = Flask(__name__)
 app.config['SECRET_KEY'] = "your-secret-key"
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///database.db'
+app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql+pymysql://root:admin123@127.0.0.1:3306/flask-crud'
 
 login_manager = LoginManager()
 db.init_app(app)
@@ -27,7 +28,7 @@ def login ():
 
   if username and password:
     user = User.query.filter_by(username=username).first()
-    if user and user.password == password:
+    if user and bcrypt.checkpw(str.encode(password), str.encode(user.password)):
       login_user(user)
       print( current_user.is_authenticated)
       
@@ -48,7 +49,8 @@ def create_user():
   username = data.get("username")
   password = data.get("password")
   if username and password:
-   user = User(username=username, password=password)
+   hashed_password = bcrypt.hashpw(str.encode(password), bcrypt.gensalt())
+   user = User(username=username, password=hashed_password)
    db.session.add(user)
    db.session.commit()
    return jsonify({"message": "Usuário cadastrado com sucesso!"})
@@ -66,10 +68,12 @@ def read_user(id_user):
 @app.route("/user/<int:id_user>", methods=['PUT'])
 @login_required
 def update_user(id_user):
-  if current_user.id != id_user:
-    return jsonify({"error": "Voce nao tem permissao para alterar esse usuário"})
   data = request.json
   user = User.query.get(id_user)
+
+  if id_user != current_user.id and current_user.role == 'user':
+    return jsonify({"error": "Voce nao tem permissao para alterar esse usuário"}), 403
+  
   if user and data.get("password"):
     user.password = data.get("password")
     db.session.commit()
@@ -80,6 +84,9 @@ def update_user(id_user):
 @login_required
 def delete_user(id_user):
   user = User.query.get(id_user)
+
+  if current_user.role != 'admin':
+    return jsonify({"message": "Operacao nao permitida"})
 
   if user and id_user != current_user.id:
     db.session.delete(user)
